add order history for admin
