+++ src/main/java/com/Ecommerce/acme/model/Product.java
@@ -25,7 +25,7 @@ public class Product {
 	private String picture;
 	
 	@ManyToOne(fetch = FetchType.LAZY)
-	@JoinColumn(name = "id_category", referencedColumnName= "id_category", insertable = false, updatable = false)
+	@JoinColumn(name = "id_category", referencedColumnName= "id_category", insertable = true, updatable = false)
 	private Category category;
 	
 }
