+++ src/main/java/com/Ecommerce/acme/validator/UserValidator.java
@@ -5,7 +5,6 @@ import org.springframework.stereotype.Component;
 import org.springframework.validation.Errors;
 import org.springframework.validation.ValidationUtils;
 import org.springframework.validation.Validator;
-
 import com.Ecommerce.acme.model.User;
 import com.Ecommerce.acme.service.UserService;
