+++ src/main/resources/templates/order.html
@@ -4,7 +4,7 @@
 
 <head>
 <meta charset="UTF-8">
-<title>Produits</title>
+<title>Historique de commande</title>
 <link th:href="@{/css/bootstrap.min.css}" rel="stylesheet" />
 </head>
