+++ src/main/resources/templates/profil.html
@@ -78,7 +78,7 @@
 								alt="avatar" class="rounded-circle img-fluid"
 								style="width: 150px;">
 							<h5 class="my-3" th:text="${person.first_name}"></h5>
-							<p class="text-muted mb-1">Commercial chez ACME</p>
+							<p class="text-muted mb-1">Admin chez ACME</p>
 							<p class="text-muted mb-4">Zone d'exploitation E.U</p>
 						</div>
 					</div>
